+++ lib/clamp/attribute/declaration.rb
@@ -25,6 +25,7 @@ module Clamp
       end
 
       def define_default_for(attribute)
+        return false if attribute.default_value.nil?
         define_method(attribute.default_method) do
           attribute.default_value
         end